Moved aligned allocation logic to detail header.
